Reduce but don't eliminate excess extern "C".
G_BEGIN_DECLS is almost never needed. Esp. around entire files.
It is for "the OS" and interoperation with some code that is still C, like profilers.

MONO_BEGIN_DECLS is often used in internal headers, in which case the same, but
not addressed here.

MONO_BEGIN_DECLS is often used in public headers, along with comprehensive
MONO_API, in which case it is redundant. This would be useful if/when mono
goes back to using .def files, which are a good idea for multiple reasons.


Commit migrated from https://github.com/mono/mono/commit/399264331b399deeb448b285eba6f39e037d98bb