Small step toward reduction/elimination of JIT icall hashing. (mono/mono#14149)
Ultimate goal is to significantly reduce or maybe completely
eliminate the hashing of JIT icalls.

It is not a small change.

This is a small digestable albeit tedious step toward that.

The "names" of opcode emulation icalls are needed for
token pasting and/or struct fields. They are needed without quotes.

This PR removes the quotes.
And then for now, just always restores them.

This is extracted from https://github.com/mono/mono/pull/14047
as a step toward making it easier to review.

Commit migrated from https://github.com/mono/mono/commit/800701fe3f370a3f1ab79723192fb3262d1b8375