Fix devirtualization in crossgen2
In some casesi that are only expressible in IL, devirtualization is
choosing a wrong virtual method. This causes a failure of the
JIT\Regression\JitBlue\GitHub_19222 test.

The test uses the following class hierarchy:
class A
{
    public virtual void f() {}
}

class B : A
{
    public override void f() {}
}

class C : B
{
    public new void f() {}
}

class D : C
{
    public new void f() {}
}

Then there is a local of type C to which we assign a reference to D.
After that the following virtual call is made:

callvirt instance int32 modopt([mscorlib]System.Runtime.CompilerServices.IsLong) A::f()

Devirtualization will choose C.f to call instead of the D.f.

This change disables devirtualization in such cases. Old crossgen does
the same thing.