Tweak codegen for Regex's FindFirstChar (#1996)
* Avoid extra field loads / comparisons in FindFirstChar

If we've already compared the current pos to the end pos as part of a minimum length check, we don't need to repeat it again later.  We can also read from locals rather than fields more than we are to save redundant field reads.

* Reduce branches in Boyer-Moore negative switch table

It's often the case that lots of negative branches have the same value.  We can coalesce them.

* Address PR feedback