Reduce execution time of RegexCharacterSetTests (#2194)
I recently wrote a large, exhaustive set of tests to validate our handling of regex character sets.  Unfortunately it ends up being too much for our CI system.  We were validating ~440 different character sets, for both RegexOptions.None and RegexOptions.Compiled, and for each validating all 65,536 character inputs... that's ~58M regex matches.  In most cases locally it was taking around 15s, but on a loaded, underwhelming CI machine, it was taking minutes in some cases.

I've reduced the load (primarily by only validating the specified characters that should be included or excluded in cases where we're already testing both the positive and negative variants of the same set), while trying to keep a reasonable semblence of the coverage.

I also made it outerloop, and moved a few other longer-running regex tests to outerloop as well.