Local live builds not picking up runtime config (#2093)
* Local live builds not picking up runtime config

The local live builds were not properly picking up the runtime
configuration varaible. This meant that libraries always attempted to
build against a Debug CoreCLR even if the developer specified
`-runtimeConfiguration Release`

* PR Feedback:

- Ensure that `$(RuntimeConfiguration)` is always set
- Remove duplicated logic in Subsets.props

* Remove unnecessary properties

* Progress

* Mix configurations

Change the live / live jobs so they mix release and debug
configurations. This will help ensure we avoid future breaks in this
area

* Undo MonoOsGorup delete

* Correct the import order

* Publish binary logs

* Pass -ci as this is a CI build

* Progress

* Test

* Set ArtifactsObjDir early enough for Mono

Mono currently relies on ArtifactsObjDir when setting
BaseIntermediateOutputPath. Not setting it leads to restore issues.

* Don't overwrite the properties

Co-authored-by: Viktor Hofer <viktor.hofer@microsoft.com>