Remove Daily Builds section (#32057)
This section has been out of date since the repo consolidation. The Master column is from November - the last build from core-setup. The other columns don't make sense for this repo, since we don't service those releases from dotnet/runtime.

We can bring this table back when we get a way to keep the links up to date. See #32055 for tracking that.