Special-case two set loops in RegexCompiler (#32003)
* Optimize .* in Singleline mode

When RegexOptions.Singleline is specified, `.*` goes from meaning `[^\n]*` to meaning `[everything]*`.  As such, when we encounter it, we can just jump to the end without actually comparing anything.  This is worth optimizing in in the compiler because, well, why not, but we don't in the interpreter because it's not a common-enough pattern to spend cycles checking for.

* Use IndexOfAny to implement set loops of 2 or 3 negated chars

* Apply suggestions from code review

Co-Authored-By: Dan Moseley <danmose@microsoft.com>

Co-authored-by: Dan Moseley <danmose@microsoft.com>