Fix OverflowException in JsonSerializer.Serialize (#1308)
* Add failing test

At first I tried to create five internal references to CustomClassToExceedMaxBufferSize and create 25 levels deep instances, but the tests ran forever. So I changed it to have several big string properties, which lead to the System.OverflowException using 3 level deep instance.

* Initial implementation. Still needs to throw the appropriated Exception

* Validate new buffer size against MaxByteArrayLength

* When newSize overflows because growBy == _renteBuffer.Length, it tries to use growBy = sizeHint
If it still overflow, we set the newSize to int.MaxValue, which will throw an OutOfMemoryException when calling ArrayPool.Rent.

* Change to make assembly smaller and reduce a extra check in the common case

* Name conventions
OuterLoop
Make the test simpler

* Improve assembly size
Run test on 64 bits only

* Added Platform Specific attribute.
Separate OuterLoop attribute, to be consistent with the other tests in this solution.