HttpClientMiniStressTest: Disable parallelism (#34941)
* Update HttpClientMiniStressTest.cs

Currently, it's possible for 4 different instances of SingleClient_ManyGets_Sync to run in parallel, each using up to 2 x 5000 sockets, totaling in 40,000 sockets. way beyond the original intent of testing 5000 client instances. since this additional stress is random in nature, I believe it is not intentional.
This change make sure that those tests will not run in parallel, this avoids some issues I've encountered in some systems.

* HttpClientMiniStressTest: Disable Parallelism: Used CollectionAttribute on the base class